generalizing parser
At this point I can define an array in the parser and grab the data
from the xml. I think I can generalize this method still further to
input the doc and the tag array, but that's still coming. I need to get
to the data and stop messing with the loader.